Updates and Fixes

+++ MacPatch Server/Server/conf/scripts/proxy/MPProxyUninstall.sh
@@ -2,9 +2,9 @@
 
 #-----------------------------------------
 # MacPatch Proxy Server Uninstall Script
-# MacPatch Version 2.1.x
+# MacPatch Version 2.7.x
 #
-# Script Ver. 1.0.0
+# Script Ver. 1.1.0
 #
 #-----------------------------------------
 clear
@@ -13,54 +13,39 @@ MP_BASE="/Library/MacPatch"
 MP_SRV_BASE="/Library/MacPatch/Server"
 MP_SRV_CONF="${MP_SRV_BASE}/conf"
 
-function checkHostConfig () {
-    if [ "`whoami`" != "root" ] ; then   # If not root user,
-       # Run this script again as root
-       echo
-       echo "You must be an admin user to run this script."
-       echo "Please re-run the script using sudo."
-       echo
-       exit 1;
-    fi
-    
-    osVer=`sw_vers -productVersion | cut -d . -f 2`
-    if [ "$osVer" -le "6" ]; then
-        echo "System is not running Mac OS X 10.7 or higher. Setup can not continue."
-        exit 1
-    fi
-}
-
-# -----------------------------------
-# CheckConfig
-# -----------------------------------
-
-checkHostConfig
+if [ "`whoami`" != "root" ] ; then   # If not root user,
+   # Run this script again as root
+   echo
+   echo "You must be an admin user to run this script."
+   echo "Please re-run the script using sudo."
+   echo
+   exit 1;
+fi
 
 # -----------------------------------
 # Run Script
 # -----------------------------------
 
 echo "This script will remove all of the files in ${MP_SRV_BASE}."
-echo "Are you sure you want to continue [Y/N]?"
-while read inputline
-do
-    answer="$inputline"
-    if [ -z "${answer}" ]; then
-        echo "Answer?"
-    else
-		if [ "${answer}" == "Y" -o "${answer}" == "y" ]; then
-        	break
-        else
-        	echo "Uninstall will not continue..."
-        	exit 0;
-    	fi	
-    fi
-done
+read -p "Are you sure you want to continue? [N] " rmAnswer
+rmAnswer=${rmAnswer:-N}
+
+if [ "${rmAnswer}" == "N" -o "${rmAnswer}" == "n" ]; then
+    echo "Uninstall will not continue..."
+    exit 0;
+fi
+
+read -p "Do you want to remove the software and patch content? [N] " swAnswer
+swAnswer=${swAnswer:-N}
 
 # Shutdown services
-$MP_SRV_CONF/scripts/StartServices.sh -u
+$MP_SRV_CONF/scripts/proxy/MPProxyConfig.py --services All --action stop
 
 # Remove MacPatch
 echo "Removing ${MP_SRV_BASE}"
 rm -rf "${MP_SRV_BASE}"
 
+if [ "${swAnswer}" == "Y" -o "${swAnswer}" == "y" ]; then
+    echo "Removing ${MP_BASE}/Content"
+    rm -rf "${MP_BASE}/Content"
+fi
